Update 10-3.cgel
Changed instances of "of" to Comp, following discussion with Brandon. Reattached "for service of notice..." higher at "prescribed." 

In terms of "in an action in the same court," I'm having trouble understanding the actual meaning of this sentence in context, causing problems for accurate annotation. Can we discuss what this sentence is actually trying to say to better understand attachment/possible ambiguity?